# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: BBang.Zip API
  version: 0.2.0
servers:
  - url: http://localhost:8081/api/v1
  - url: http://localhost:8082/api/v1
  - url: http://localhost:8083/api/v1
  - url: http://localhost:8084/api/v1
paths:
  /auth/check-username:
    get:
      summary: 아이디 중복 확인
      parameters:
        - in: query
          name: username
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
...
  /orders/{orderId}:
    get:
      summary: 주문 단건
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
...
  /payments/{paymentId}:
    get:
      summary: 결제 상태 조회
      parameters:
        - in: path
          name: paymentId
          required: true
          schema: { type: string }
      responses:
        '200': { description: ok }
...